add ManagementSupport to JMS, VFS and Mail transports updated axis2 version to 1.4-SNAPSHOT and its dependencies
git-svn-id: https://svn.eu.apache.org/repos/asf/synapse/trunk/java@648315 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/transports/src/main/java/org/apache/synapse/transport/base/MetricsCollector.java
@@ -89,7 +89,7 @@ public class MetricsCollector {
         timeoutsReceiving++;
     }
 
-    public synchronized void incrementBytesReceived(int size) {
+    public synchronized void incrementBytesReceived(long size) {
         bytesReceived += size;
     }
 
@@ -105,7 +105,7 @@ public class MetricsCollector {
         timeoutsSending++;
     }
 
-    public synchronized void incrementBytesSent(int size) {
+    public synchronized void incrementBytesSent(long size) {
         bytesSent += size;
     }
 }
